Create worker to forward information to a remote server.
Information forwarded so far are: creation/deletion of sales and
creation and deletion of 'inv_movements' which are changes on
inventories caused either by a sale or inventory transfer.

The goal is to be able to record inventory and display sale's reporting
on the cloud, so there are less dependency on local server, making it
easier to replace one.